Fix issues with YAML include tag.
Calling load_yaml_config_file(config_path) causes issues when trying to
load from multiple files in a directory since config_path is modified in
the closure.

This also separates the parse code block into its own function so that
there is no restriction that the object be a dictionary for included
files.  This means that scalars, lists, etc. can also be stored in
separate files -- for example, the scene component expects a list.